change id, health card number, and contactnumber from int to long

+++ Hospital System/src/main/java/Controllers/Admin/GetBestStrategies/GetBestByOperationIncome.java
@@ -7,7 +7,7 @@ public class GetBestByOperationIncome implements GetBestStrategy {
     StaffManaging sm = new StaffManager();
 
     @Override
-    public int findBest() {
+    public long findBest() {
         return sm.getBestStaffByMostOperationIncome();
     }
 }
